Move tomorrowio base entity to separate module
