+++ homeassistant/components/tomorrowio/__init__.py
@@ -3,7 +3,6 @@
 from __future__ import annotations
 
 from pytomorrowio import TomorrowioV4
-from pytomorrowio.const import CURRENT
 
 from homeassistant.components.sensor import DOMAIN as SENSOR_DOMAIN
 from homeassistant.components.weather import DOMAIN as WEATHER_DOMAIN
@@ -11,10 +10,8 @@ from homeassistant.config_entries import ConfigEntry
 from homeassistant.const import CONF_API_KEY
 from homeassistant.core import HomeAssistant
 from homeassistant.helpers.aiohttp_client import async_get_clientsession
-from homeassistant.helpers.device_registry import DeviceEntryType, DeviceInfo
-from homeassistant.helpers.update_coordinator import CoordinatorEntity
 
-from .const import ATTRIBUTION, DOMAIN, INTEGRATION_NAME
+from .const import DOMAIN
 from .coordinator import TomorrowioDataUpdateCoordinator
 
 PLATFORMS = [SENSOR_DOMAIN, WEATHER_DOMAIN]
@@ -57,35 +54,3 @@ async def async_unload_entry(hass: HomeAssistant, config_entry: ConfigEntry) -> 
             hass.data.pop(DOMAIN)
 
     return unload_ok
-
-
-class TomorrowioEntity(CoordinatorEntity[TomorrowioDataUpdateCoordinator]):
-    """Base Tomorrow.io Entity."""
-
-    _attr_attribution = ATTRIBUTION
-    _attr_has_entity_name = True
-
-    def __init__(
-        self,
-        config_entry: ConfigEntry,
-        coordinator: TomorrowioDataUpdateCoordinator,
-        api_version: int,
-    ) -> None:
-        """Initialize Tomorrow.io Entity."""
-        super().__init__(coordinator)
-        self.api_version = api_version
-        self._config_entry = config_entry
-        self._attr_device_info = DeviceInfo(
-            identifiers={(DOMAIN, self._config_entry.data[CONF_API_KEY])},
-            manufacturer=INTEGRATION_NAME,
-            sw_version=f"v{self.api_version}",
-            entry_type=DeviceEntryType.SERVICE,
-        )
-
-    def _get_current_property(self, property_name: str) -> int | str | float | None:
-        """Get property from current conditions.
-
-        Used for V4 API.
-        """
-        entry_id = self._config_entry.entry_id
-        return self.coordinator.data[entry_id].get(CURRENT, {}).get(property_name)
