+++ homeassistant/components/tomorrowio/weather.py
@@ -37,7 +37,6 @@ from homeassistant.helpers.entity_platform import AddEntitiesCallback
 from homeassistant.helpers.sun import is_up
 from homeassistant.util import dt as dt_util
 
-from . import TomorrowioEntity
 from .const import (
     CLEAR_CONDITIONS,
     CONDITIONS,
@@ -61,6 +60,7 @@ from .const import (
     TMRW_ATTR_WIND_SPEED,
 )
 from .coordinator import TomorrowioDataUpdateCoordinator
+from .entity import TomorrowioEntity
 
 
 async def async_setup_entry(
